refactor: decouple legacy/workspace mixin
